FlaskBE/requirements.txt
flask
PyGithub
openai
requests
gunicorn